Created patient search on homepage, and built the session-based links that allow someone logged into this module to jump back and forth between this and rwandaprimarycare.
diagnosisFindPatient.jsp is built on core's find patient js object, but the resultsHandler method is overwritten to only do something if the search string matches the patient's ID.
This has the effect of making the patient search only sensitive to either typing or scanning (preferrably scanning) a barcode.

For search by name, etc..., you *could lookup a patient in the registration module, and then jump across.